- added more javadocs

+++ src/main/java/Controller/addProductController.java
@@ -7,13 +7,11 @@ import javafx.collections.transformation.FilteredList;
 import javafx.collections.transformation.SortedList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Alert;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 /** Controller class "addProductController" provides logic for addProductView */
@@ -129,10 +127,35 @@ public class addProductController implements Initializable {
         }
     }
 
+    @FXML
     public void onProductRemoveButtonClicked() {
+        // Get the selected part from the associated parts table
         Part selectedPart = productPartTable.getSelectionModel().getSelectedItem();
-        newProduct.deleteAssociatedPart(selectedPart);
+
+        if (selectedPart != null) {
+            // Show confirmation dialog
+            Alert confirmationAlert = new Alert(Alert.AlertType.CONFIRMATION);
+            confirmationAlert.setTitle("Remove Associated Part");
+            confirmationAlert.setHeaderText(null);
+            confirmationAlert.setContentText("Are you sure you want to remove this associated part?");
+            Optional<ButtonType> result = confirmationAlert.showAndWait();
+
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                // Remove the part if the user confirms
+                newProduct.deleteAssociatedPart(selectedPart);
+                productPartTable.setItems(newProduct.getAllAssociatedParts());
+                System.out.println("Associated part removed: " + selectedPart.getName());
+            }
+        } else {
+            // Show an error alert if no part is selected
+            Alert errorAlert = new Alert(Alert.AlertType.ERROR);
+            errorAlert.setTitle("No Part Selected");
+            errorAlert.setHeaderText(null);
+            errorAlert.setContentText("Please select a part to remove.");
+            errorAlert.showAndWait();
+        }
     }
+
 
     public void onAddProductExitClicked (){
         stage = (Stage) addProductPane.getScene().getWindow();
